The Status entity was removed
It was considered to eliminate this entity because it added more complexity to the database design and also took up more memory space.

Therefore, it was decided to use the term "soft-delete" and add an IsDeleted property of type bool in the necessary entities. In case you need to obtain the name of the status, that is, if it is ACTIVE or INACTIVE, you can use the GetStatusName method, which is an extension of the ModelWithSoftDelete class.