Separate seeding data from the entity configuration
The benefit of this is that can group the methods that create seed data into a single block to be executed only in the development environment. In addition, code repetition is also avoided. So there is no need to use the 'WebHostEnvironment' class in the configuration class of each entity.