refactor: Removed coupling between chatbot and appointment module
The following members have been added to the "Shared.Appointments" layer:
* CreateAppointmentRequest
* AvailableTimeRangeRequest
* AvailableTimeRangeResponse

Two interfaces have been created:
* IGetAvailableHoursUseCase
* ICreateAppointmentUseCase

These interfaces act as an "intermediary" or "bridge" between the chabot and the Appointment module.

It was also possible to add the implementation of the use case (e.g. CreateAppointmentUseCase) in the shared layer, however, those internal details belong to the "Appointments" module (they are its own details).