refactor: Rename namespace to CPlugin.Net
This change has been made to facilitate reuse.
"CPlugin.Net" directory contains everything related to the plugin loader and can be distributed in a NuGet package to reuse the code in other projects.

"PluginAttribute" class can also be distributed in a NuGet package using "CPlugin.Net.Contracts" as namespace. This way plugins do not need to reference the plugin loader.

In other words, "CPlugin.Net" would be a new project that would be in another repository.