Create strongly typed objects to encapsulate error and success messages
This change has only been applied for messages that have placeholders. In fact, formatted messages I don't use it that often.

The benefit of this approach is that the consumer does not need to know how the message is formatted, in my opinion, it makes the code look more expressive rather than using string.Format directly.
The consumer does not even need to know what string.Format means or how it works.