fix: Add TimeSpanJsonConverter type
This fixes this error: The JSON value could not be converted to System.TimeSpan.
This error is occurring in the use case to cancel appointments by the dentist. For example, when the client sends in the "startHour" field a value like "07:38", this error is generated.
This error can also occur in other use cases where time slots are handled.